Refactor non-view-related code into new `core` module Support calling exports without a request
